requirements.txt
-e .[avatar]
colour-runner==0.0.4
coverage==4.2.0
dj-database-url==0.4.1
dj-inmemorystorage==2.1.0
factory-boy==2.8.1
flake8==2.5.4
flake8-import-order==0.7
incuna-test-utils==8.0.0
pillow>3.3.2
psycopg2-binary==2.8.5
